♻️ Move validateCertificate and verifyLiveness to certificate_generation - Move and update all the related tests accordingly
